removes cache and extracts to two helpers
